fix(remix-react): fix infinite LiveReload loop on Firefox
