+++ integration/bug-report-test.ts
@@ -1,15 +1,11 @@
 import { test, expect } from "@playwright/test";
-import { spawn } from "cross-spawn";
-import getPort from "get-port";
-import * as readline from "node:readline";
-import fs from "fs";
-import path from "path";
 
 import { PlaywrightFixture } from "./helpers/playwright-fixture";
-import { createFixtureProject } from "./helpers/create-fixture";
-import { js } from "./helpers/create-fixture";
+import type { Fixture, AppFixture } from "./helpers/create-fixture";
+import { createAppFixture, createFixture, js } from "./helpers/create-fixture";
 
-let appFixture: Awaited<ReturnType<typeof startDevServer>>;
+let fixture: Fixture;
+let appFixture: AppFixture;
 
 ////////////////////////////////////////////////////////////////////////////////
 // 💿 👋 Hola! It's me, Dora the Remix Disc, I'm here to help you write a great
@@ -44,138 +40,67 @@ let appFixture: Awaited<ReturnType<typeof startDevServer>>;
 //    ```
 ////////////////////////////////////////////////////////////////////////////////
 
-async function startDevServer(fixture: {
-  files?: { [filename: string]: string };
-}) {
-  let port = await getPort();
-  let serverUrl = `http://localhost:${port}`;
-
-  let projectDir = await createFixtureProject(fixture);
-
-  let watchProcess = spawn(
-    "node",
-    ["node_modules/@remix-run/dev/dist/cli.js", "dev"],
-    {
-      cwd: projectDir,
-      env: {
-        ...process.env,
-        PORT: port.toString(),
-      },
-    }
-  );
-
-  await new Promise<void>((isReady, onError) => {
-    let rl = readline.createInterface(watchProcess.stdout);
-    let timeout = setTimeout(() => {
-      onError(new Error("Timed out waiting for watch server"));
-    }, 10000);
-
-    rl.on("line", (input) => {
-      if (input.match(/Remix App Server started at http:\/\/localhost/)) {
-        rl.removeAllListeners();
-        clearTimeout(timeout);
-        isReady();
-      }
-    });
-    rl.on("close", () => {
-      clearTimeout(timeout);
-      onError(new Error("Watch server did not start correctly"));
-    });
-  });
-
-  let waitForRebuild = (timeout: number = 10000) => {
-    let rl = readline.createInterface(watchProcess.stdout);
-    return new Promise<void>((resolve, reject) => {
-      let timeoutRef = setTimeout(() => {
-        reject(new Error("Timed out waiting for watch server rebuild"));
-      }, timeout);
-
-      rl.on("line", (input) => {
-        if (input.match(/Rebuilt in/)) {
-          rl.removeAllListeners();
-          clearTimeout(timeoutRef);
-          resolve();
-        }
-      });
-      rl.on("close", () => {
-        clearTimeout(timeoutRef);
-        reject(new Error("Watch server died unexpectedly"));
-      });
-    });
-  };
-
-  return {
-    projectDir,
-    serverUrl,
-    waitForRebuild,
-    close: async () => {
-      if (!watchProcess.kill()) {
-        watchProcess.kill(9);
-      }
-    },
-  };
-}
-
-test.beforeEach(async () => {
-  appFixture = await startDevServer({
+test.beforeAll(async () => {
+  fixture = await createFixture({
+    ////////////////////////////////////////////////////////////////////////////
+    // 💿 Next, add files to this object, just like files in a real app,
+    // `createFixture` will make an app and run your tests against it.
+    ////////////////////////////////////////////////////////////////////////////
     files: {
       "app/routes/index.jsx": js`
         import { json } from "@remix-run/node";
+        import { useLoaderData, Link } from "@remix-run/react";
 
-        export async function loader() {
-          return json({});
+        export function loader() {
+          return json("pizza");
         }
 
         export default function Index() {
-          return "initial";
+          let data = useLoaderData();
+          return (
+            <div>
+              {data}
+              <Link to="/burgers">Other Route</Link>
+            </div>
+          )
+        }
+      `,
+
+      "app/routes/burgers.jsx": js`
+        export default function Index() {
+          return <div>cheeseburger</div>;
         }
       `,
     },
   });
+
+  // This creates an interactive app using puppeteer.
+  appFixture = await createAppFixture(fixture);
 });
 
-test.afterEach(() => appFixture.close());
+test.afterAll(() => appFixture.close());
 
 ////////////////////////////////////////////////////////////////////////////////
 // 💿 Almost done, now write your failing test case(s) down here Make sure to
 // add a good description for what you expect Remix to do 👇🏽
 ////////////////////////////////////////////////////////////////////////////////
 
-test("after file changes the app reloads if the loader is slow", async ({
-  page,
-}) => {
+test("[description of what you expect it to do]", async ({ page }) => {
   let app = new PlaywrightFixture(appFixture, page);
+  // You can test any request your app might get using `fixture`.
+  let response = await fixture.requestDocument("/");
+  expect(await response.text()).toMatch("pizza");
 
+  // If you need to test interactivity use the `app`
   await app.goto("/");
-  expect(await app.getHtml()).toMatch("initial");
+  await app.clickLink("/burgers");
+  expect(await app.getHtml()).toMatch("cheeseburger");
 
-  await Promise.all([
-    fs.promises.writeFile(
-      path.join(appFixture.projectDir, "app", "routes", "index.jsx"),
-      js`
-        import { json } from "@remix-run/node";
+  // If you're not sure what's going on, you can "poke" the app, it'll
+  // automatically open up in your browser for 20 seconds, so be quick!
+  // await app.poke(20);
 
-        export async function loader() {
-          await new Promise((resolve) => {
-            setTimeout(resolve, 2000);
-          });
-
-          return json({});
-        }
-
-        export default function Index() {
-          return (
-            <div id="changed">changed</div>
-          );
-        }
-      `
-    ),
-    appFixture.waitForRebuild(),
-    app.page.waitForNavigation(),
-    app.page.waitForSelector("#changed"),
-  ]);
-
-  expect(await app.getHtml()).toMatch("changed");
+  // Go check out the other tests to see what else you can do.
 });
 
 ////////////////////////////////////////////////////////////////////////////////
